The HSTWCS method 'wcs2header' was modified to append the SIP and IDC distortion model keywords to the header using the '.update()' method instead of extending the header with a CardList.  This corrected a problem with the CardList not fully updating the original header making the appended cards unaccessible using the dictionary lookup syntax.
The logic for using the 'readModel()' method of HSTWCS was also improved to not try and read the IDCModel when no 'idctab' had been specified in the header.  

Finally, a small bug was fixed in the 'wcscorr' module for recognizing what rows are available for being updated so that the update will actually occur. 
 


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/stwcs@12392 fe389314-cf27-0410-b35b-8c050e845b92